Added logic for the insurance costs and fixed database structure.
Got most of the concepts figured out for querying for objects but was running into issues with making the queries play nice with the return values I wanted.

Next steps are to get to the bottom  of the issues with my queries that are causing issues with the database.

Then moving on to front-end.